WORKING some operations, most registers
This is mostly a working model and I think if fleshed out would continue
to improve. However, I think that I need a way to abstract similar
operations to reduce code duplication. Example, all the 8bit load
operations are fundementally the same with a different first operand

LD A and LD B will be the same in the guts, just pointing to a different
8 bit registers. Let's see if I can get that working.